+++ source/textUtils/segFlag.py
@@ -9,6 +9,7 @@ from enum import IntFlag
 _AUTO: int = 1 << 0
 _UNISCRIBE: int = 1 << 1
 _CHINESE: int = 1 << 2
+_ICU: int = 1 << 3
 
 
 class CharSegFlag(IntFlag):
@@ -26,3 +27,4 @@ class WordSegFlag(IntFlag):
 	AUTO = _AUTO
 	UNISCRIBE = _UNISCRIBE
 	CHINESE = _CHINESE
+	ICU = _ICU